refactor: modified fields from Integer to Long

+++ src/main/java/com/claudioscagliotti/thesis/dto/response/MovieDto.java
@@ -11,6 +11,6 @@ public record MovieDto(String overview,
                        List<GenreEnum> genreIds,
                        String posterPath,
                        String releaseDate,
-                       Integer tmdbId
+                       Long tmdbId
 ) {
 }
